requirements.txt
pdf2image==1.17.0
pdfplumber==0.11.7
pytesseract==0.3.13
python-dotenv==1.1.1
streamlit==1.48.0
google.generativeai==0.8.5